updated versions

+++ lab4/src/lab4localization/Navigation.java
@@ -87,7 +87,7 @@ public class Navigation {
 	 * travelTo function which takes as arguments the x and y position in cm Will travel to designated position, while
 	 * constantly updating it's heading
 	 */
-	public void travelTo(double x, double y, float speed) {
+	public void travelTo(double x, double y, int speed) {
 		double minAng;
 		while (Math.abs(x - odometer.getX()) > CM_ERR || Math.abs(y - odometer.getY()) > CM_ERR) {
 			minAng = (Math.atan2(y - odometer.getY(), x - odometer.getX())) * (180.0 / Math.PI);
